extract colour conversion functions into another module
